Add back in fixture mk_tmp_dirs
