+++ jwst/pipeline/tests/helpers.py
@@ -10,6 +10,7 @@ import tempfile
 # simply to make available from this module.
 from ...tests.helpers import (
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
@@ -21,20 +22,6 @@ SCRIPT_DATA_PATH = path.join(SCRIPT_PATH, 'data')
 
 
 @pytest.fixture
-def mk_tmp_dirs():
-    """Setup testing directories"""
-    tmp_current_path = tempfile.mkdtemp()
-    tmp_data_path = tempfile.mkdtemp()
-    tmp_config_path = tempfile.mkdtemp()
-
-    old_path = os.getcwd()
-    try:
-        os.chdir(tmp_current_path)
-        yield (tmp_current_path, tmp_data_path, tmp_config_path)
-    finally:
-        os.chdir(old_path)
-
-
 def update_asn_basedir(asn_file, root=None):
     """Create an association with filenames update
     for a different directory
