+++ jwst/pipeline/tests/test_calwebb_image2.py
@@ -6,6 +6,7 @@ import pytest
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
 )
