+++ jwst/pipeline/tests/test_calwebb_spec2.py
@@ -8,6 +8,7 @@ from shutil import copyfile
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
 )
 
 from ...associations.asn_from_list import asn_from_list
