+++ jwst/pipeline/tests/test_calwebb_spec2_nrs_msa.py
@@ -7,6 +7,7 @@ from shutil import copy as file_copy
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
