+++ jwst/pipeline/tests/test_calwebb_spec3_mir_mrs.py
@@ -14,6 +14,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
     update_asn_basedir,
