+++ jwst/pipeline/tests/test_calwebb_spec3_mir_mrs_dither.py
@@ -12,6 +12,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
