+++ jwst/pipeline/tests/test_calwebb_spec3_nrs_msa.py
@@ -8,6 +8,7 @@ from shutil import copy as file_copy
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
