+++ jwst/pipeline/tests/test_nis_ami3.py
@@ -6,6 +6,7 @@ from os import path
 
 from .helpers import (
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
     update_asn_basedir,
