+++ jwst/pipeline/tests/test_nrc_coron3.py
@@ -7,6 +7,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
     update_asn_basedir,
