+++ jwst/pipeline/tests/test_nrc_detector1.py
@@ -6,6 +6,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
