+++ jwst/pipeline/tests/test_nrc_image3.py
@@ -4,6 +4,7 @@ from os import path
 
 from .helpers import (
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
