+++ jwst/pipeline/tests/test_nrc_tso3.py
@@ -8,6 +8,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
     update_asn_basedir,
