+++ jwst/pipeline/tests/test_nrs_spec2_bright_obj.py
@@ -6,6 +6,7 @@ from os import path
 from .helpers import (
     SCRIPT_DATA_PATH,
     abspath,
+    mk_tmp_dirs,
     require_bigdata,
     runslow,
 )
